Crate translator_common

Re-exports

pub use proto::TranslatorClient;

Modules

proto

Structs

TranslatorEntrypoint

Stub with a manual implementation of oak::WasmEntrypoint, to be used by other nodes that want to instantiate a translator node. The manual implementation must be kept in sync with the auto-generated implementation from the translator module.

Functions

translate